requirements.txt
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
flask-paginate==2022.1.8
Flask-WTF==1.0.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
Pillow==9.0.1
pygame==2.1.2
pytesseract==0.3.8
python-decouple==3.6
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.1
